requirements.txt
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
whatsapp-api-client-python
python-dotenv
requests
termcolor
unidecode
openai
yt-dlp
ffmpeg-python
pandas 